exercise-15/requirements.txt
kafka-python==2.0.2
streamlit==1.40.0
pandas==2.1.4
matplotlib==3.8.2
plotly==5.18.0
watchdog==3.0.0